requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
github3.py==4.0.1
tweepy==4.14.0
praw==7.7.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
